backend/requirements.txt
Django==5.2.9
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-decouple==3.8
supabase==2.24.0
